Created a mercurial repository for the 'memory allocation' projects group.
In November 2008 I created 3 projects with the intention to conduct
microbenchmarks that would show the difference in performance between native
C++, C++/CLI running inside a CLR console application and C++/CLI running in a
Windows Forms application when allocating large arrays and filling them. Later I
conducted similar microbenchmarks in various configurations of Java running on
Sun's JRE.

This repository is intended to track the source code of these 3 projects each in
a different branch.